# agendly-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Agendly API",
        "description": "REST API for Agendly platform — scheduling, students, services and appointments management.",
        "termsOfService": "https://agendly.com.br/terms",
        "contact": {
            "name": "Agendly Support",
            "url": "https://agendly.com.br",
            "email": "support@agendly.com.br"
        },
        "license": {
            "name": "Proprietary",
            "url": "https://agendly.com.br/license"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Local development"
        },
        {
            "url": "https://staging.api.agendly.com.br",
            "description": "Staging environment"
        },
        {
            "url": "https://api.agendly.com.br",
            "description": "Production environment"
        }
    ],
    "paths": {
        "/v1/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register new user",
                "description": "Creates a new user and returns an authentication token.",
                "operationId": "d9e61e050fc948da6eb592172578662d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Lucas Fajardo"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "lucas@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123456"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|XyzTokenExampleHere"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/v1/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User login",
                "description": "Authenticates user and returns a token.",
                "operationId": "7abb047ced1329103f0767160e852fc9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "lucas@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|XyzTokenExampleHere"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/v1/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout user",
                "description": "Revokes the current authentication token.",
                "operationId": "fb4913f284432ecf8c1394ad6c99a14e",
                "responses": {
                    "200": {
                        "description": "Logged out successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logged out successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/students": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "List students",
                "description": "Returns all students linked to the authenticated personal trainer.",
                "operationId": "a76b73bf8e79114f2b2729f244665d86",
                "responses": {
                    "200": {
                        "description": "Students retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Student"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Student": {
                "title": "Student",
                "description": "Represents a student/client linked to a personal trainer.",
                "required": [
                    "id",
                    "personal_id",
                    "name",
                    "status",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "description": "Unique student identifier",
                        "type": "integer",
                        "example": 1
                    },
                    "personal_id": {
                        "description": "Identifier of the personal trainer responsible for the student",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Full name of the student",
                        "type": "string",
                        "example": "Carlos Silva"
                    },
                    "email": {
                        "description": "Student email address",
                        "type": "string",
                        "format": "email",
                        "example": "carlos@example.com",
                        "nullable": true
                    },
                    "phone": {
                        "description": "Phone number including area code",
                        "type": "string",
                        "example": "11999999999",
                        "nullable": true
                    },
                    "status": {
                        "description": "Current status of the student",
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive"
                        ],
                        "example": "active"
                    },
                    "notes": {
                        "description": "Medical notes or additional observations",
                        "type": "string",
                        "example": "Mild right knee injury",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-02-13T21:31:41.000000Z"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-02-13T21:31:41.000000Z"
                    },
                    "deleted_at": {
                        "description": "Soft deletion timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Lucas Fajardo"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "lucas@email.com"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Bearer token authentication. Example: 'Authorization: Bearer {token}'",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Authentication operations"
        },
        {
            "name": "Students",
            "description": "Operations related to students"
        }
    ]
}